extract the date into a separate file
