+++ modules/logic.js
@@ -1,10 +1,25 @@
 /* eslint-disable no-restricted-globals */
-/* eslint-disable import/no-unresolved */
-/* eslint-disable import/prefer-default-export */
+export const container = document.querySelector('.container');
+export const contact = document.querySelector('#contact');
+export const displayContact = document.querySelector('#display_contact');
+export const displayBooks = document.querySelector('#display_list');
+export const displayForm = document.querySelector('#display_form');
+export const bookText = document.getElementById('book-text');
 
-import {
-  container, contact, displayContact, displayBooks, displayForm, bookText,
-} from './modules/data.js';
+const date = document.querySelector('.date');
+
+setInterval(() => {
+  date.innerHTML = `${Date().slice(0, 24)}`;
+}, 1000);
+
+contact.innerHTML = `
+  <h2> Contact information </h2>
+
+  <ul>
+    <li>Email Address: Kennediowusu@gmail.com</li>
+    <li>Phone Number: +233546985015</li>
+    <li>Street Address: Kwame Nkrumah Downside</li>
+  </ul>`;
 
 const showOnlyContact = () => {
   contact.style.display = 'block';
